Update C3 module (#1705)
